Move run function to a dedicated go file
