+++ pkg/core/engine/run.go
@@ -1,0 +1,53 @@
+package engine
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/sirupsen/logrus"
+)
+
+// Run runs the full process for one manifest
+func (e *Engine) Run() (err error) {
+
+	logrus.Infof("\n\n%s\n", strings.Repeat("+", len("Pipeline")+4))
+	logrus.Infof("+ %s +\n", strings.ToTitle("Pipeline"))
+	logrus.Infof("%s\n\n", strings.Repeat("+", len("Pipeline")+4))
+
+	for _, pipeline := range e.Pipelines {
+
+		err := pipeline.Run()
+
+		e.Reports = append(e.Reports, pipeline.Report)
+
+		if err != nil {
+			logrus.Printf("Pipeline %q failed\n", pipeline.Title)
+			logrus.Printf("Skipping due to:\n\t%s\n", err)
+			continue
+		}
+	}
+
+	err = e.Reports.Show()
+	if err != nil {
+		return err
+	}
+	totalSuccessPipeline, totalChangedAppliedPipeline, totalFailedPipeline, totalSkippedPipeline := e.Reports.Summary()
+
+	totalPipeline := totalSuccessPipeline + totalChangedAppliedPipeline + totalFailedPipeline + totalSkippedPipeline
+
+	logrus.Infof("Run Summary")
+	logrus.Infof("===========\n")
+	logrus.Infof("Pipeline(s) run:")
+	logrus.Infof("  * Changed:\t%d", totalChangedAppliedPipeline)
+	logrus.Infof("  * Failed:\t%d", totalFailedPipeline)
+	logrus.Infof("  * Skipped:\t%d", totalSkippedPipeline)
+	logrus.Infof("  * Succeeded:\t%d", totalSuccessPipeline)
+	logrus.Infof("  * Total:\t%d", totalPipeline)
+
+	// Exit on error if at least one pipeline failed
+	if totalFailedPipeline > 0 {
+		return fmt.Errorf("%d over %d pipeline failed", totalFailedPipeline, totalPipeline)
+	}
+
+	return err
+}
